Triggers dialog (7.0-dev): Fixed an issue where instant switch profile becomes NULL after entering triggers dialog for another profile.
Oops, a logic error: until now, NVDA assumed that the cleared checkmark for instatn switch profile meant the instant switch flag could be nullified. This was observed after chekcing a profile as an instant switch profile, moves to a different profile and opens triggers dialog without selecting OK from the main add-ons dialog. This has been fixed (was a major issue since the introduciton of the triggers dialog).