6.4: A major issue with profile deletion and inconsistent instant switch profile index is now gone.
Consider the following:
1. Four or more profiles are in use, with position 3 and 4 occupied by the currently active and the instant switch profiles respectively.
2. User delets profile at postion 2.
Expected: When switching back, position 2 profile should become active.
actual: Due to flaw in the prevProfile storage, position 3 becomes active (instnat switch comes back).
To prevent this, in add-on 6.x and 7.0, a warning will be given when attempting to delete a profile while an instant switch profile is active (to be fixed in a later 7.x release). This is such an important issue that requires immediate release.