7.0 beta: allow some beta testers to switch to stable channel for updates (will be removed in official release, to be resurrected before LTS hits the air).
The update qualify check now takes an optional cv (current version) string, which is used to let NVDA obtain older releases (this will be used in 7.x LTS/8.0). Also adds a preliminary control to select add-on update channel.
If stable channel is selected, the stable URL will be looked up first before resorting to connecting to the development URL.