Top and bottom notification (8.0-dev/LTS): Play a beep if a broadcaster reaches top and bottom of playlists.
For now, IAccessibleChildID will be checked to play alert beeps (also added a handler for up and down arrows). If SPL API doesn't improve responsiveness, the timer will be removed (if it does, then SPL API will be used).
Due to testing required, this will be reserved for LTS/8.0-dev.