7.0 beta 5 (project purple): It is no longer required to stay in the playlist viewer in order to invoke SPL Assistant.
Until now, broadcasters were required to stay in the playlsit viewer in order to invoke SPL Assistant. Although this provided protection (in that it allowed SPL Assistant command to pass through in other places), it meant users couldn't obtain crucial info such as automation status when not focused in playlist viewer (besides, scripts for JAWS does not have this restriction). Effective immediately, this restriction is gone.

Specificlaly:
* One can now invoke SPL Assistatn from anywhere as long as one is focused on a window that belongs to Studio (exceptions include Track Tool and SPL Recorder).
* To send SPL Assistant command to OS, one must press this command twice or more.
This is destined for add-on 7.0 and later as add-on 6.x is on its last legs.