Welcome dialog (8.0-dev/LTS): Present a welcome dialog similar to how NVDA Core does when ran for the first time.
To introduce new users to the operation of the Studio add-on, a welcome dialog has been added that presents add-on highlights, layer commands and more. So far, it is a work in progress (it'll not be complete without adding buttons to read add-on readme and such). Also added a command (Alt+NVDA+F1) to open this dialog in case it was dismissed.
Due to work involved and to receive feedback from people, this is destined for add-on 8 and 7.x LTS.