7.1: Changed prev profile storage format.
Instead of storing index to the previously active profile, the actual profile name will be used. This should prevent the index confusion error seen in earlier commits where nVDA thought it was still using instant switch profile when it should not. Consequently, profile deletion will be permitted when instant switch profile is active (to be implemented once testing shows it is working).