7.1: InstantProfile string is no longer kept in the config dictionary at runtime.
Suppose a user resets add-on settings. Because instant profile is gone yet the string is recorded in the config, it causes confusion. Because InstantProfile is meant as an offline flag to be invoked next time the add-on runs (to search for the instant switch profile), it isn't necessary to keep this string around at runtime, thus remove it.