7.1: When encoder settings file becomes corrupt, a message will be presented when the add-on loads.
Same issue as add-on settings file corruption: ConfigObj raises parse error if a file appears to be corrupted. Although it would have been great to handle this problem as the encoder file is loaded, it has issues with user interface, namely an error dialog popping up inthe midst of a broadcast and so on. Thus an error dialog will be presented during app module initialization if encoder settings file is indeed corrupted.
This is destined for add-on 7.1.