Track comments (8.0-dev/7.x LTS): Do not allow comments to be set or announced for hour markers and friends.
Thankfully, hour markers, break notes and so on have no file names, thus it is possible to filter this out. Also, when checking for comments, the existence of a given filename, plus whether it is NULL or not will be checked, reducing code duplication.
Added translator comments (reminded by Derek Riemer's post regarding this).