7.2 and beyond: Remove really old update data routines, remove the possibility of caching old-style config keys.
Various 'package' update data such as size and date are no longer in use, so remove code that detects them.
Also, as a continuation on parting ways with the old config format, do not cache old-style config keys. This results in some space savings for the config cache.
These changes are destined for 7.2 and later.