Columns Explorer R2 (8.0-dev/partial LTS): Columns Explorer script has been modified to allow Contorl+NvDA+number row commands to work.
In 7.x days, just checking to see if just a number was pressed was enough. This is no longer the case, as one would use Control+NvDA+number row to invoke columns explorer. Thus a routine to strip plus signs away and obtian just the digit has been added to this script to take care of this. This routine will be gone in add-on 9.0.
This is destined for 8.0 and LtS.