Columns Explorer @r/Track Tool (8.0-dev): Foundation to support Columns Explorer in track Tool and a dialog to configure this (partially operational).
A new key has been added to config dictionary to house Columns Explorer TT list. This list will be consulted from Track Tool to locate needed columns. Also, the configuration dialog is there but it is partially operational - as in not taken advantage of by Track Tool yet (that is next).
This is destined for add-on 8.0 and later only.