Columns Explorer R2/Track Tool (8.0-dev): Track Tool track item now honors Columns Explorer list.
Same routine as main app module except header will be consulted - a primitive form of indexOf function. This means no more old scripts for individual properties such as overlap and others.
Also fixed an issue where NvDA was looking for colexprange when it is unnecessary, which prompted error tones to be heard when attempting to reset settings.
This is destined for add-on 8.0 and later. only.