Track Tool (8.0-dev/LTS): Performance improvements when obtaining column data when pressing Control+NVDA+1 through 0.
One of the causes of performance regressions was use of linear column header lookup. Given that previous work on column data fetcher on Studio indicate that internal column positions will not change, one could use a constant data structure (a tuple, for instance) to store column header names for each position (done in this commit). This speeds up column header lookup significantly.
This will be available in one form or another in 8.0 and LTS.