Config cache (8.0-dev): Completely rewritten cache config and config save routines.
With the old single-level keys gone, just cache the entire dictionary. This makes the cache routines easier to read and follow.
Also, in cache config routine, never convert included columns key to list, as it causes the cache and the config pool version to become out of sync quickly and causes the profile to be saved to disk when no modifications were mode.
This is destined for add-on 8.0 and later.