7.3: Vast performance improvements when fetching objects in SPL Assistant scripts.
According to Jamie Teh (NV Access), it is faster to use obj.getChild instead of obj.children[index] when retrieving a specific child object (reason being the children is built each time it is called in different MainLoop cycle). Tests show this is indeed the case: average of eight times speed-up when using obj.getChild(index). Although this isn't significant for users, it is a significant performance improvement when fetching objects. However, because fetching objects the first time takes a while, caching will still occur.
This is destined for add-on 7.3 and later.