7.4 or later: Allow broadcasters to configure auto update interval.
Because stable versions are released less frequently, it would be safe to assume that users would check for updates once a week. Thus introduce a new setting in advanced options dialog that'll let users specify update interval between 1 and 30 days. The overall switch is still auto update checkbox.
This is destined for add-on 7.4 and later.