SPLConfig is now powered by Config Hub.
The new ConfigHub structure, powered by ChaimMap, allows multiple profiles to be listed as a list of dictionaries. As this is now a class, it allows runtime information to be carried along, including active profile, switch history and others.
Specifically:
* Various module-level functions were transfered to the new ConfigHub object, including profile fetcher, init code and others.
* InitConfig and SaveConfig now calls object-level functions.
* The ConfigHub constructor now performs both file loading and unlock routines, as well as initializing various attributes.
* Profile switching was simplified. Instead of copying data around, it just swaps dictionaries around.
* Due to referencing nature of Python, profile cache uses copy.deepcopy.
* ConfigHub also includes profile management features, including dedicated methods to create, cache, rename and delete profiles.
For backward compatibility names of functions in the config module will be kept until 9.0 (internally they call functions in the ConfigHub object). This is destined for add-on 8.0 and later and cannot be backported to LTS.