Emergency 7.3 r2: make sure to handle both GitHub and old URL's.
Because GitHub uses cloud to store files, the URL returned is a bit misleading. Also, because public snapshots are stored on the add-on server, handling one URL format means discarding the other, so be sure to handle both.
Specifically: when checking for file name, instead of taking the filename part of the add-on name, it'll split at the identifier name.
This is an emergency release of add-on 7.3.