7.4: Progress dialog calling routine reordered, update size removed.
Borrowing from NVDA Core: be sure to start progress dialog before using another thread to run update check and the resulting dialog. This menas a private reference to progress dialog will be kept by splupdate module.
As update file size is no longer consulted, remove this variable from update  check routine and from the update module.
This is destined for add-on 7.4 or later.