7.4: Officially change display name for the add-on, do not allow global commands to work in secure mode.
As noted during recent reviews for add-ons, it might be helpful to detect one is in secure mode, thus disabling global features. For SPL, it concerns SPL Controller entry and SPL window focus switch commands. Specifically:
* In secure screens, one cannot obtain handle to SPL window at all, so these comands will not work properly (at least SPL Controller will say handle is 0 when in fact SPL is running).
* Allowing folks to switch to Studio window in secure mode is a security risk, in that one can jump out of an important prompt (not even possible in secure screens).
Thus use globalVars.appArgs.secure to make sure this is enforced.
Also took this time to (finally) give proper label for the add-on (Station Playlist Studio -> StationPlaylist Studio).
These changes are destined for add-on 7.4 and later.