7.4/8.0/LTS: Simplify update check by checking for version differences only.
As noted by Jamie Teh (NV Access): it is better to check for string differences rather than making it complicated. Thus starting with 7.4, update check will check version differences. This means the error message that says one is using a newer version of the add-on is no more, as well as simplifying LTS check process.
This is destined for add-on 7.4.