7.4: only sedlect channels if different from the default.
As the update channel variable defaults to the branch name, it is safe to assume that this will be consulted as the default channel if everything else fails. For 7.4, it is stable, for LTS, LTS, for 8.x, stable/dev/beta. This will not apply to try builds (staging branch has URL already set).