16.10.1/15.2: In some cases, after finding a track, users were not able to interact with the found track in some systems.
A backport of a fix in master branch (vertical column navigation): after finding a track, without telling Studio where the broadcaster is, it'll argue with NVDA as to which track it is dealing with. Thus forcefully tell Studio to focus to the right track (press SPACE twice).
This is destined for add-on 16.10.1 and 15.2 LTS.