Vertical column navigation (17.1-dev): Introduce ability to configure vertical column announcement combo box.
The ultimate idea of vertical column navigation is not just to move to next or previous track and announce the current column one is reviewing; the goal is to let this be configurable. Thus, a new combo box is now available to allow users to configure which columns should be announced when using vertical column navigation commands. By default, it'll review the current column, and the rest corresponds to Studio 5.10 columns. Due to possible confusion, this will be a global setting.
This is destined for add-on 17.1. Report focus routine for track item will be modified to hande this new option.