Vertical column navigation (17.1-dev): REport focus will now take vertical column announcement setting into account when announcing track information.
Officially bring vertical column announcement setting to life. Column 0 (Status) will be announced if the setting is set to 'Status', otherwise an index for the currently selected column will be retrieved. All this will not happen if user tells NVDA to announce whichever column one is reviewing (default value).
This is destined for add-on 17.1-dev.