Vertical column navigation (17.1-dev): Take care of display order changes, announce status if navigating by columns other than status column.
Again, make sure to take care of display order changes, as there might be a miscommunication between NVDA and Studio as to which column is which. This is done by passing in header string to column content announcer when vertical column navigation is active.
Also, if looking at columns other than status column, announce status along with the column information to help broadcasters get an idea as to if a track is selected or not. A new boolean flag will be passed in that'll control this output (in column announcer), and if yes, track name and a blank space will be prepended just prior to header/info string, and an empty string otherwise. This flag is not set by default (the only time this flag will be set is when doing vertical column navigation).
These enhancements are destined for add-on 17.1.