Vertical column navigation: it is now possible to move vertical by pressing Control+Alt+up/down arrow keys. fixes #10.
One can now use Control+Alt+up or down arrow keys to move vertically up or down columns. By default, the column one is reviewing will be used, but one can set a specific column to be reviewed irresepctive of the column one is currently on.
This is destined for add-on 17.1.