GUI Helper (17.1-dev): Use GUI Helper services provided by NVDA 2016.4 to make the UI presentation consistent with NVDA, with Alarms Center as a testbed.
In NVDA 2016.4, GUI Helper services were added that abstracts layout rendering away from vairous dialogs in hopes of making UI presentation consistent throughout NVDA. As an effort to do this for add-ons, GUI Helper will be used.
As stage 1, Alarms Center has been updated to use GUI Helper if available. This results in edit field for alarm settings moving to the right of the alarm prompt, with OK and Cancel buttons aligning to the right.
This change is destined for add-on 17.1. A full switchover to GUI Helper will happen later in 2017.