16.12: Use Studio 5.20 API to locate scheduled for and track starts in values.
Studio 5.20 API provides a way to locate scheduled for (track starts) and track starts in value (in case of the latter, it'll take up to a second for the new value to appear via the API). This provides huge performance improvements compared to object navigation emulation.
To avoid regressions, this will be available in 16.12 stable only.