16.12/15.4-LTS: In Studio 5.20, cart insert status is announced.
Cart insert status is also shown on the same text box as cart edit mode status (in Studio 5.20 API, it's a separate wparam), thus handle this case. This also means verbosity check cannot be done for cart edit/insert mode status (a necessary sacrifice unless a better way is found later).
This is destined for 16.12 and 15.4-LTS.