Track item column getter: fetch column inclusion set only once to prevent calling getitem method in ConfigHub multiple times.
Consider the following: display order is turned off, custom column order is set. When this happens, description pieces for track items will be built, but this means column inclusion set will be consulted every time a header is looked up. This causes ConfigHub.__getitem__ to be called multiple times (up to seventeen times, once per header). Thus store this set once in order to prevent this from happening. This optimization doesn't make sense if IAccessible builds description pieces, but provides some speedups if the above scenario is in use (particularly on slow computers).
This optimization is destined for 17.1-dev and later.