GUi Helper (17.1-dev): converted column announcement dialog.
For column announcement dialog, it was found that it's gotten taller because the label for the list box (column announcement order) is above the list box itself, not to the left. This will necessitate moving move up/down buttons to the right of the list box (otherwise there is a huge whitespace). The button arrangement will be done once NVDA 2016.4 is released.