Metadata streaming gatherer (17.1-dev): optimized metadata streaming status gatherers and config checkboxes.
Perhaps read the API wrong: when metadata URL is turned off 9including DSP encoder), it returns '0' not '-1'. This opens up rooms for optimizations, including:
* App module: metadata status scripts will now check the return value of the function instead of coercing a negative value to zero via a variable. This avoids creating a temp variable and improves readability.
* Metadata streaming checkboxes: Checkbox.SetValue can now work with return values directly instead of using a coerced view.
Also, when announcing overall metadata streaming status (on demand or when Studio starts), conditional list comprehension is used to build the flags for URL's (does not affect DSP encoder, which also was simplified by using the returned value from the SendMessage function). This speeds things, as it eliminates unnecessary for loop (comprehension does this automaticlaly) and condition checking can be included in list comprehension call.
Also, for presentation, the list of URL's will be built directly instead of kept in a variable.
These optimizations are destined for add-on 17.1.