GUi Helper services (17.1-dev): Reorganized Column Announcements checkboxes and metadata streaming checkboxes.
For column announcement checkboxes, only one list wil be used to use checkboxes instead of three. Also, instead of using a variable, the newly added checkbox will be appended to the checkboxes collection.
For metadata streaming dialog, checkboxes will be appended to the streams liss.
For Columns Explorer, label text is really a constant, so interpolate while constructing the combo boxes.