SPL Controller/5.20 support (17.1-dev): introduce a new command to obtain status information.
A new command (SPL Controller, Q) is now availalbe in SPL Controller layer to announce status information. This include playback status, automation, microphone, line-in, record to file, cart edit and insert modes, same info as SPL Studio's status bar. The first two are available for all versions, while the others are reserved for Studio 5.20 and later.
This is destined for add-on 17.1.