Library scan monitor improvements, including use of Studio 5.10 API to detect library scan completion. re #15
In the old days, when there was no easy way of telling when a library scan was finished, one assumed that temporary scan count will not change in the next few seconds. It turns out this is flawed:
* For scans that involve thousands of tracks (for example, when scanning an entire drive), scan count will remain the same for a while before further scans are done. This meant that one would terminate library scan monitor too early, giving potentially misleading information about items scanned.
* In Studio 5.10 and later, there is an easy way to tell when the library scan has completed, thus making reliance on scan count unappealing.
Thus changed the library scan monitor routine to use Studio 5.10 API (this means breaking backwards compatibility with Studio 5.0x, which means this fix is destined for add-on 17.1).