Code cleanup: variables that are really constants or generate constants and then used in a subsequent function clals are now truly constants.
In some cases, expressions such as:
someConst = constantGenExpression
someFunc(someConst)
isn't quite necessary. This is a bit better:
someFunc(constGenExpression)
results in less bytecode and improves fetching variables that are really constants (this is the case when the generated constant is used in one place only). This observation has been applied to SPL Controller, library scan initiator in SPL Assistant and in Track Tool app module.