17.01: Use timestamps to announce Cart Explorer reentry reminder. re #19.
It is better to provide a more accurate picture by saying that Cart Explorer should be entered again whenever timestamps changes (when Cart Edit is off and if carts were modified, SPL saves cart files to disk). Although it introduces a nagging reminder, it is better this way as broadcasters should reenter this mode to see updated cart assignments. This also lays the foundation for refreshing cart information automaticlaly (later in 2017).
This is destined for add-on 17.01 and 15.5-LTS.