Playlist snapshots (17.1-dev): Add playlist snapshots config key to ConfigHub as a set.
For compact representation, it is best to add playlist snapshots flags as a set (similar to 'included columns' key), and now this set has been included as part of COnfigHub. In the app module, this set will be looked up as a last resort of no flags are specified. The next commit will be the user-visible config dialog.