Playlist snapshots (17.1-dev): Use a dedicated section for playlist snapshots variables, now booleans.
A conscious decision due to inconsistency: dialogs that are branches off from add-on settings sets keys that are in different section, so why was it that playlist snapshots did not follow this rule? Now corrected.
Also, for ease of readability and maintenance, playlist snapshots keys are now booleans 9or would be integers later). The caveat is that playlist snapshots function in the pap module must convert dictionaries to lists (via list comprehension), a necessary thing but acceptable for now.