Update check (17.1-dev): Use a regular expression to locate new version string.
A better solution than splitting strings: because the file name for the add-on package will stay the same for a long time, just use regular expressions to locate the new version string (stationplaylist-[version].nvda-addon). This improves update version check routine and is resistant to URL scheme changes.
This is destined for add-on 17.04.