Status info (17.1-dev): Check if Studio is running before invoking this command.
SPL Controller, Q: if a custom gesture has been defined for this command, the script will assume that Studio is running (but it might not be running at all). In case Studio isn't running, catch this and report it (same routine as other SP lUtils commands where custom gestures can be defined).