Code cleanup: Renamed 'statusAPI' to 'studioAPI' to better reflect its purpose, use hardcoded constant for user32.WM_USER.
In previous versions of the add-on, statusAPI did what it is supposed to do - obtain status information. However, with the advent of metadata streaming and other features, it became clear that this function needed to be renamed to something else. Because this function is a thin wrapper around user32.SendMessage with the handle to Studio and message type filled in, and because this function is now used to set some options, it is more appropriate to call it 'studioAPI'.
Also, for global plugin, WM_USER will be hardcoded (1024 or 0x400) for consistency reasons (other places uses the constant directly, while the global plugin used variable names, causing consistency headahce).
These changes are destined for add-on 17.04 and later.