Playlist snapshots (17.04-dev): Allow custom commands to be defined for playlist snapshtos script, some presentation tweaks.
It is now possible for broadcasters to assign a hotkey to invoke playlist snapshots. Pressing once will speak and braille brief snapshot info, while pressing it twice will present the full snapshots display. To prevent mishaps with multiple HTML windows appearing and slowing down performance, the command will not work if pressed more than twice.
Also tweaked how snapshot info is presented. Specificlaly, because the top artist and others will be announced when the script is invoked once, counters will be fetched early.
All this work is destined for add-on 17.04.