SPL gestures: do not allow built-in and custom SPL add-on commands to proceed if handle to studio is not present.
In studio Demo, before the main window appears, registration screen is shown. At this time, main window handle to Studio isn't present, which causes commands such as remainig time of the track, temperature, playlist snapshots and others to fail (not announcing anything, playing error tones, or giving wrong info). This has been mitigated via a new function that'll announce an error message if this is such a case.
This fix is destined for add-on 17.04.