Vertical column nav optimization: let column tracker be managed by the overlay class instead of exposing this in the app module.
in the app module, column nav tracker (the variable used to track which column a broadcaster is reviewing) should be handled by the overlay class. This prevents someone from changing column nav behavior by modifying the app module and allows the object (track item) itself to record which column one is consulting.
This is destined for add-on 17.04. A similar change will be made to track Tool app module.