Code deduplication/optimization: just use one copy of layer helper.
Code duplication: layer helper is defined in both the app module and the global pulgin, and they are identical. Thus it is better to let the app module use the global plugin (not the other way around as it causes import error, and if fooled to ignore nonexist SPL window, causes bytecode to be generated when the app isn't running) in order to let just one version deal with both Controller and Assistant layers.