Playlist snapshots: shortest track is recognized correctly if the first track of a plyalist is the shortest track of them all. re #22.
Reported by a broadcaster and noted as a logic error: just because shortest track segue is None doesn't mean it should be assigned to the first segue, thus causing inequality to fail (the inequality checks for current min, which could very well be the very first segue). Also this takes care of a problem where all tracks in a playlistlist have the same length.