Debugging framework (17.04/17.2-dev): initial foundation for debugging framework with a debug logging printer. re #24.
In NVDA 2017.1, it is possible to allow NVDA to provide debug logging for a single session. Thus take advantage of this by introducing a basic debugging framework that'll let the add-on output important debugging information (such as command execution steps) to the log for analysis later.
An initial solution will be in place in 17.04, with functionality coming in via subsequent stable releases, full picture to be ready by 17.2-dev (summer/fall).