Cart Explorer 3 (17.04): refresh carts via the init function in order to reduce loop traversals.
Optimization: In add-on 17.01, when refreshing cart banks, cart files were read up to two times (two for loops). The best case was only one loop (no refresh needed). To reduce loop traversal, combine aspects of cart explorer init and refresh functions into cart explorer init that'll also be able to refresh carts if the timestamps for carts exists (error conditions to be taken care of later). This optimization results in code reuse and allows just one function to provide debug output, and in some cases only one loop will be run (check the timestamp, and if they are equal for all, don't update carts by processing cart files), as well as saving the add-on from processing all carts unnecessarily.
This is destined for add-on 17.04.