17.05: NVDA will stop the time-based profile trigger end timer if the Studio app module exits before time expires. re #30.
Consider the following scenario: a time-based profile is active and the Studio app exits prior to the expiration of the profile. This results in attribute error and other exceptions. Catch this problem by checking to make sure the trigger end timer (the function that'll switch from a time-based profile back to another) is stopped, and if it's still running, forcibly stop this.
This is destined for 17.05 and 15.7-LTS.