17.06/15.8-LTS: Track Finder will no longer raise attribute erro if search forward or backward from the last or first track, respectively. re #33.
A technical oversight: when search forward or backward, focused track.next/.previous is chosen, respectively. In case of the last or first track, it'll result in None being selecte,d which causes attribute error to be raised. Because the internal track finder function has a special case for it, fool the internal function to using None object, which allows NVDA to pop up search error dialog.
This is destined for add-on 17.06 and 15.8-LTS.