Search history (17.2-dev): move the just entered search text to the top of the search history. re #27.
Previously, the search history combo box would show a list of search terms as it is stored in the find text array. However, in order to provide a true history-like fell, items in this array must be moved around (most recent term is on top). Thus provide this check if necessary.