Next/current track announcer (17.2-dev): include internal player name to which tracks will be played or is playing. re #34.
Requested by a broadcaster: in certain situations, two separate streams or mixers are fed to two internal players in Studio. Thus allow NVDA to announce which player will be playing the next track and/or the player that's playing the current track. By default, this is turned off, but can be turned on from settings dialog (next commit).
This is destined for add-on 17.08.