SPL Controller (17.2-dev): for status info script, look up status bar messages from Studio for versions earlier than 5.20.
For Studio 5.11 and earlier, use status bar fetcher (part of the app module) to locate the status bar messages and present it to the user when SPL Controller, Q is pressed. This foundation will also be used in the next commit where broadcasters will be able to obtain current track title via SPL Controller.