Update check (17.08): prevent integer/interval overflow by checking for update interval only. re #36.
In the past, when update checker ran, it would also run a timer if continuous checking argument is set to true. If this happens and if interval is greater than 24 days, this results in an integer overflow. Resolved this by checking for interval alone and no longer calling the update check timer automaticlaly. This is only for automatic update checks.