Minimal and debugger: honor NVDA's minimal startu pflag by not announcing the sign-on message.
If 'nvda -m' is specified, globalVars.appArgs.minimal flag is set. If so, do not announce SPL Studio version message at startup. This is useful for situations where fast startup and readiness is essential.
This is destined for 17.09 and provides one of the foundations for improved debugging framework, which will include unit tests and what not.