UI overhaul: introduce a non-functioning General Settings button.
As part of replaceing dialogs with list-based sizers, a new 'General add-on settings' dialog will be created. This will result in less clutter on the add-on settings dialog and better settings organization. ultimately, buttons to open dialogs (except profile controls) will be replaced by a categories list and a sizer to pull in relevant options together.
This work is destined for some time later this year or 2018.