UI overhaul: introduce General add-on settings dialog.
The general add-on settings will house some settings that are currently part of the main add-on settings window. The following settings can be changed from this screen:
* Beep announcement
* Message verbosity
* Braille timer
* Library scan announcement
* Track categories
* Top and bottom of playlist notification
* Hour announcement
* Track comment announcements
* Requests window notification
* Vertical column navigation (technically should be part of column announcements)
At the moment tne general settings will pull in settings from SPLConfig. For demo purposes, options from the main dialog will be used, with a commented out section that does use SPLConfig for the next stage in UI overhaul (categories list). Tests show this dialog is functioning.