Playlist remainder (17.09/15.9-LTS): say the playlist is not ready when it is in fact not ready.
In previous versions, when performing SPL Assistant, D, NVDA will say '00:00' when a playlist is not ready (not loaded). Thus catch this and provide more accurate message. Also, if the playlist is ready yet the focused object is the tracks list, move down to the first track and continue with reminder calculation.

# Conflicts:
#	addon/appModules/splstudio/__init__.py