Volatile config (17.10): avoid caching profiles if volatileConfig flag is on.
If config is volatile, there's no need to cache them in order to find out if they need to be saved, as the only file system call is 'read'. Therefore cacheConfig function will return early if volatileConfig flag is turned on, usually used when one wishes to have a read-only config for various reasons 9among them troubleshooting issues with options).