Microphone alarm timer (17.11/15.10): add a new function to cause Studio app module to restart mic alarm timer. re #38.
If two profiles have differing mic alarm values and if profiles are switched, the time will run using the old value. Thus allow the timer to be restarted.
At the moment there is really  no elegant way to notify the pap module to do this (coming later in 2017 on the stable releases). However, this can be done by using a new function in splmisc module that'll call the app module function on behalf of SPLConfig.
This is destined for both stable and LTS as it affects all Studio releases.