Mic alarm timer (17.11/15.10): reworded doExtraAction function to check mic alarm and act appropriately based on options. re #38.
The actual culprit of alarm timer problem is the fact that mic alarm was assumed to be on when in fact it might not. As a result, when switching profiles, the mic alarm/alarm interval threads will run until mic status has changed. Thus check if mic alarm must be on, and if not, turn off these alarm threads. Also, look for a possible refactor where mic alarm thread termination routines can be separated into its own function in order to fix a possibly related problem where the mic alarm threads may run after Studio exits.