Extension points/Action (17.12): introduce extensionPoints.Action. re #40.
In NVDA 2017.4, one can specify extension points and be notified of things happening such as actions, profile switches and what not. In SPL add-on, several actions are defined in a new module named splactions, including app ready, profile loaded/saved, broadcast profile switch and others. To maximize compatibility, extension points will be used if running on NVDA 2017.4 or later.