17.11/15.10-LTS: never touch status of instant switch profile when deleting the active profile.
Consider the following: three profiles are define, with one of the being an active profile and another being an instant switch profile. When the active profile other than normal profile is gone, instant switch flag is nullified, caused by a lax 'or' statement in delete button handler in add-on settings. Catch a specific case for instant switch profile, which means if an active profile is gone, instant switch profile will not be affected.