Metadata connector (17.11): just call the connector function from config UI.
Consider the following: when a broadcaster clicks 'OK' button from add-on settings, as part of the OK button handler, metadata connector will be called. First, a handle to Studio is retrieved, and if it returns something, it'll be passed onto the connector function. The connector itself will go through two conditionals: the handle checker, and a double-check to make sure the Studio is really alive. In this case, this results in the handle going through three conditionals, which isn't ideal (the fastest case is if Studio is not alive yet or is dying, in which case the connector will not be called; but for most cases, add-on settings would be active while Studio is alive).
One way to optimize this is just calling splmisc.metadataConnector with no regard to status of Studio. In this case, this results in two checks. Also, because the connector will be called from other modules (such as the main app module file) with no regard to presence of Studio window handle, it made more sense to let the handle be retrieved straight from the connector function whose job is to place metadata streaming flags by calling Stdio API which does require handle to Studio window.
This optimization is destined for 17.11.