Track comments (17.11/15.10): catch end of file exception.
Reported by a user: the app module didn't load completely as track comments pickle file raised EOF (end of file) error when parsing, a likelyhood of a corruption. For now, let it pass, but it'll raise an error in a future commit that'll require broadcasters to set up comments again.