Extension points (17.12): let config dialog notify config switch extension point in the add-on. re #40.
When profiles are switched 9as in leaving the config dialog with a different profile selected) or when a profile is deleted, make sure to let metadata streaming handle server switches and let the app module restart mic alarm timer. At the moment metadata streaming server switching is working, and the next commit will extend this to app module action registrant.